+++ CHANGELOG.md
@@ -7,4 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Project releases
 
+## v1.1.0 - 18 August 2025 - First version
+
+#### Added
+
+-   Tool to truncate video with
+
+#### Changed
+
+-   Tabs system to navigate between tools
+
 ## v1.0.0 - 04 August 2025 - First version
+
+#### Added
+
+-   Tool to convert video into images